+++ lib/src/ui/pages/cupertino/list_tile.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'list_tile.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$tileConfigHash() => r'32a167a27482e070c76570bf9db86e29dba18eea';
+
+/// See also [TileConfig].
+@ProviderFor(TileConfig)
+final tileConfigProvider =
+    AutoDisposeNotifierProvider<TileConfig, ListTileConfig>.internal(
+  TileConfig.new,
+  name: r'tileConfigProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$tileConfigHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$TileConfig = AutoDisposeNotifier<ListTileConfig>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member