+++ lib/src/ui/pages/cupertino/widgets.dart
@@ -1,9 +1,11 @@
 import 'package:dreambook/src/ui/pages/shared/shared_code_view.dart';
 
+import 'list_tile.dart';
 import 'slider.dart';
 import 'switch.dart';
 
 final cupertinoItems = <CodeItem>[
   switchItem,
   sliderItem,
+  listTileItem,
 ];