+++ lib/src/ui/theme.dart
@@ -1,13 +1,22 @@
 import 'package:dreambook/src/utils/kv.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'theme.g.dart';
 
 ThemeData appTheme([Brightness brightness = Brightness.light]) {
+  const primary = Colors.cyan;
+  const cupertinoPrimary = CupertinoColors.systemTeal;
   return ThemeData(
     colorScheme:
-        ColorScheme.fromSeed(seedColor: Colors.cyan, brightness: brightness),
+        ColorScheme.fromSeed(seedColor: primary, brightness: brightness),
     splashFactory: InkSparkle.splashFactory,
+    cupertinoOverrideTheme: CupertinoThemeData(
+      brightness: brightness,
+      primaryColor: cupertinoPrimary,
+      applyThemeToAll: true,
+      textTheme: const CupertinoTextThemeData(primaryColor: cupertinoPrimary),
+    ),
     useMaterial3: true,
   );
 }